Use unsigned format for HDCP_STATUS in error message
Issue Type: Overflowed integer argument

Cast data.Status to (uint32_t) with %u format specifier to avoid signed-type
underflow when printing HDCP_STATUS enum value in SendResponse failure path.

Tracked-On: NIACP3-1382

+++ daemon/daemon.cpp
@@ -252,7 +252,7 @@ void HdcpDaemon::MessageResponseLoop(void)
             sts = m_SdkSocket.SendResponse(data, appId);
             if (SUCCESS != sts)
             {
-                HDCP_ASSERTMESSAGE("SendResponse failed. %d", data.Status);
+                HDCP_ASSERTMESSAGE("SendResponse failed. %u", (uint32_t)data.Status);
 
                 // If we can't communicate with the app, destroy the connection
                 PortManagerHandleAppExit(appId);
